fixed @MediaPlayerTimeChanged bug
It was because TOTAL had an extra empty entry after I removed cantab as an option. So when I tried to search for a string in an empty cell, it returned an error.
Changed the size of TOTAL when it was created in analyzeVideo.m. There should be no empty cells now and therefore no error.