Added save2Excel
Run save2Excel.m and it'll ask you to choose .m files to convert into excel. It will then ask you which folder to save to. You can make a new folder at this point. It will just save to the same name as the original .m files except the extension is now .xlsx for excel. If a file with exactly the same name already exists in that folder, it'll give you a choice to overwrite, rename, or skip.
It'll convert ONLY the TOTAL variable into excel file.
You can select multiple files and it'll convert all of them.